play-book-api/chat_service/requirements.txt
Django>=4.2.0
celery>=5.3.0
python-dotenv>=1.0.0
kombu>=5.3.0
redis>=4.5.0
amqp>=5.1.0
langchain>=0.3.25
chromadb-client>=1.0.11